requirement.txt
asgiref==3.8.1
cffi==1.17.0
cryptography==43.0.0
Django==5.1
djangorestframework==3.15.2
drf-yasg==1.21.7
ecdsa==0.19.0
inflection==0.5.1
packaging==24.1
pyasn1==0.6.0
pycparser==2.22
python-jose==3.3.0
pytz==2024.1
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
